% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_n_data.R
\docType{data}
\name{imported_data}
\alias{imported_data}
\alias{departement_simplifie}
\alias{prenom_insee}
\alias{annee_max}
\title{departement simplifie}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 96 rows and 5 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 3517156 rows and 5 columns.

An object of class \code{numeric} of length 1.
}
\usage{
departement_simplifie

prenom_insee

annee_max
}
\description{
departement simplifie

#' departement
#'
#' @name imported_data
#' @docType data
#' @references \url{https://geoservices.ign.fr/documentation/diffusion/telechargement-donnees-libres.html#admin-express}
#' @keywords data
'departement'
prenom INSEE

annee max
}
\references{
\url{https://www.data.gouv.fr/fr/datasets/contours-des-departements-francais-issus-d-openstreetmap/}

\url{https://www.data.gouv.fr/fr/datasets/ficher-des-prenoms-de-1900-a-2018/}
}
\keyword{data}
