% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{nav_}
\alias{nav_}
\title{Navigation bar (top, expand)}
\usage{
nav_(name, x)
}
\arguments{
\item{name}{character. Title.}

\item{x}{named character vector. List of navigation choice.}
}
\description{
Navigation bar (top, expand)
}
