% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{nav_item}
\alias{nav_item}
\title{Navigation bar item (top)}
\usage{
nav_item(label, id)
}
\arguments{
\item{label}{character. Link destination.}

\item{id}{character. Label.}
}
\description{
Navigation bar item (top)
}
